+++ blocklets/ai-studio/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.1.146 (2023-11-20)
+
+- feat: set template empty page
+- fix: inherits empty parameters of call prompt
+- chore: update uploader package to latest
+
 ## 0.1.145 (2023-11-20)
 
 - fix: use property name as the file name in git